Added a reset button to the config
