+++ src/main/config.js
@@ -2,7 +2,7 @@
 import { existsSync, readFileSync, writeFileSync } from "fs";
 import { app } from "electron";
 import path from 'path';
-import { mergeDeep } from '../common/util';
+import { mergeDeep, cloneDeep } from '../common/util';
 
 export const CONFIG_DEFAULT = {
     version: app.getVersion(),
@@ -180,7 +180,7 @@ export function loadConfig() {
     const config_path = path.join(app.getPath('userData'), CONFIG_FILENAME);
     if (existsSync(config_path)) {
         try {
-            config = mergeDeep(config, JSON.parse(readFileSync(config_path, { encoding: 'utf8' })));
+            config = mergeDeep(cloneDeep(config), JSON.parse(readFileSync(config_path, { encoding: 'utf8' })));
         } catch (e) { }
     }
     config.version = app.getVersion();
@@ -189,6 +189,6 @@ export function loadConfig() {
 
 export function updateConfig(new_config) {
     const config_path = path.join(app.getPath('userData'), CONFIG_FILENAME);
-    config = mergeDeep(config, new_config);
+    config = mergeDeep(cloneDeep(config), new_config);
     writeFileSync(config_path, JSON.stringify(config), { encoding: 'utf8' });
 }
