+++ src/main/modules/settings.js
@@ -1,10 +1,11 @@
 
 import { stringMatchQuality } from "../../common/util";
-import { config } from '../config';
-import { BrowserWindow } from "electron";
+import { BrowserWindow, ipcMain } from "electron";
 import { format as formatUrl } from 'url';
 import path from 'path';
 import { getTranslation, getAllTranslation } from "../../common/local/locale";
+import { config, updateConfig, CONFIG_DEFAULT } from '../config';
+import { updateModules } from '../executor';
 
 const isDevelopment = process.env.NODE_ENV !== 'production';
 
@@ -45,6 +46,37 @@ export function createSettingsWindow() {
 
     settings_window.on('close', hideWindow);
     settings_window.removeMenu();
+
+    async function updateConfigHelper(new_config) {
+        if (config.general.global_shortcut !== new_config.general.global_shortcut) {
+            try {
+                const ret = globalShortcut.register(new_config.general.global_shortcut, toggleMain);
+                if (ret) {
+                    globalShortcut.unregister(config.general.global_shortcut);
+                } else {
+                    new_config.general.global_shortcut = config.general.global_shortcut;
+                }
+            } catch (e) {
+                new_config.general.global_shortcut = config.general.global_shortcut;
+            }
+        }
+        const old_config = JSON.parse(JSON.stringify(config));
+        updateConfig(new_config);
+        await updateModules(old_config);
+    }
+    ipcMain.on('update-config', (_, new_config) => {
+        updateConfigHelper(new_config);
+    });
+    ipcMain.on('reset-config', (_) => {
+        updateConfigHelper(CONFIG_DEFAULT);
+        settings_window.webContents.send('update-config', config);
+    });
+}
+
+export function destroySettingsWindow() {
+    if (settings_window && !settings_window.isDestroyed()) {
+        settings_window.destroy();
+    }
 }
 
 const SettingsModule = {
