+++ src/renderer/settings/setting/quality-setting.js
@@ -1,5 +1,5 @@
 
-import React, { useState } from 'react';
+import React, { useState, useRef } from 'react';
 import { TextField } from '@material-ui/core';
 import { getTranslation } from '../../../common/local/locale';
 
@@ -13,7 +13,9 @@ const styles = {
 
 function QualitySetting(props) {
     const [error, setError] = useState(null);
+    const [value, setValue] = useState(null);
     const onUpdate = (e) => {
+        setValue(e.target.value);
         if (e.target.value.length > 0 && parseFloat(e.target.value) >= 0 && parseFloat(e.target.value) <= 1) {
             props.onUpdate(parseFloat(e.target.value));
             setError(null);
@@ -21,11 +23,17 @@ function QualitySetting(props) {
             setError(getTranslation(props.config, 'quality_error'));
         }
     };
+    const last_option = useRef(null);
+    if (last_option.current !== props.option) {
+        last_option.current = props.option;
+        setValue(props.option);
+        setError(null);
+    }
     return (
         <div>
             <TextField
+                value={value}
                 style={styles.text}
-                defaultValue={props.option}
                 variant="outlined"
                 type="number"
                 onChange={onUpdate}
