+++ src/renderer/settings/setting/shortcut-setting.js
@@ -1,5 +1,5 @@
 
-import React, { useState } from 'react';
+import React, { useState, useRef } from 'react';
 import { TextField } from '@material-ui/core';
 import { getTranslation } from '../../../common/local/locale';
 
@@ -23,7 +23,9 @@ function isValidShortcut(shortcut) {
 
 function ShortcutSetting(props) {
     const [error, setError] = useState(null);
+    const [value, setValue] = useState(null);
     const onUpdate = (e) => {
+        setValue(e.target.value);
         if (isValidShortcut(e.target.value)) {
             props.onUpdate(e.target.value);
             setError(null);
@@ -31,11 +33,17 @@ function ShortcutSetting(props) {
             setError(getTranslation(props.config, 'shortcut_error'));
         }
     };
+    const last_option = useRef(null);
+    if (last_option.current !== props.option) {
+        last_option.current = props.option;
+        setValue(props.option);
+        setError(null);
+    }
     return (
         <div>
             <TextField
+                value={value}
                 style={styles.text}
-                defaultValue={props.option}
                 variant="outlined"
                 onChange={onUpdate}
                 error={error !== null}
