+++ src/renderer/settings/setting/text-setting.js
@@ -1,5 +1,5 @@
 
-import React from 'react';
+import React, { useState, useRef } from 'react';
 import { TextField } from '@material-ui/core';
 
 const styles = {
@@ -11,14 +11,21 @@ const styles = {
 };
 
 function TextSetting(props) {
+    const [value, setValue] = useState(null);
     const onUpdate = (e) => {
+        setValue(e.target.value);
         props.onUpdate(e.target.value);
     };
+    const last_option = useRef(null);
+    if (last_option.current !== props.option) {
+        last_option.current = props.option;
+        setValue(props.option);
+    }
     return (
         <div>
             <TextField
+                value={value}
                 style={styles.text}
-                defaultValue={props.option}
                 variant="outlined"
                 onChange={onUpdate}
             ></TextField>
